moved generateDragImage into its own component
