+++ TransferList/TransferList.js
@@ -19,6 +19,7 @@ import {useCallback, useEffect, useRef, useState} from 'react';
 
 import Button from '../Button';
 import CheckboxItem from '../CheckboxItem';
+import CustomDragImage from './CustomDragImage';
 import Icon from '../Icon';
 import ImageItem from '../ImageItem';
 import Skinnable from '../Skinnable';
@@ -370,11 +371,15 @@ const TransferListBase = kind({
 	},
 
 	render: ({css, disabled, firstList, firstListMaxCapacity, firstListMinCapacity, height: defaultHeight, itemSize: defaultItemSize, listComponent, moveOnSpotlight, noMultipleDrag, orientation, renderImageItem, renderItem, secondList, secondListMaxCapacity, secondListMinCapacity, setFirstList, setSecondList, showSelectionOrder, verticalHeight}) => {
+		const item = document.querySelectorAll(`.${css.draggableItem}`)[0];
+		const itemBg = item?.querySelectorAll(`.${itemCss.bg}`)[0];
+		console.log(itemBg)
+
+		const [customImage, setCustomImage] = useState(CustomDragImage(listComponent, item, itemBg)());
 		const [firstListLocal, setFirstListLocal] = useState(firstList);
+		const [position, setPosition] = useState(null);
 		const [secondListLocal, setSecondListLocal] = useState(secondList);
 		const [selectedItems, setSelectedItems] = useState([]);
-		const [position, setPosition] = useState(null);
-
 		const height = ri.scaleToRem(orientation === 'horizontal' ? defaultHeight : verticalHeight);
 		const itemSize = ri.scale(defaultItemSize);
 		const width = orientation === 'horizontal' ? 'inherit' : '100%';
@@ -383,6 +388,11 @@ const TransferListBase = kind({
 		let scrollToRefFirst = useRef(null);
 		let scrollToRefSecond = useRef(null);
 
+		useEffect(() => {
+			const image = CustomDragImage(listComponent, item, itemBg)();
+			setCustomImage(image);
+		}, [item, listComponent]);
+
 		const getScrollToFirst = useCallback((scrollTo) => {
 			scrollToRefFirst.current = scrollTo;
 		}, []);
@@ -390,68 +400,6 @@ const TransferListBase = kind({
 		const getScrollToSecond = useCallback((scrollTo) => {
 			scrollToRefSecond.current = scrollTo;
 		}, []);
-
-		const setCommonElementStyles = (element) => {
-			const item = document.querySelectorAll(`.${css.draggableItem}`)[0];
-			if (item) {
-				if (listComponent === 'VirtualList') {
-					const itemBg = item.querySelectorAll(`.${itemCss.bg}`)[0];
-					element.style.backgroundColor = window.getComputedStyle(itemBg).backgroundColor;
-					element.style.borderRadius = window.getComputedStyle(itemBg).borderRadius;
-				} else {
-					element.style.backgroundColor = window.getComputedStyle(item, ':before').backgroundColor;
-					element.style.borderRadius = window.getComputedStyle(item, ':before').borderRadius;
-				}
-
-				element.style.border = '1px solid black';
-				element.style.height = item.clientHeight + 'px';
-				element.style.left = "0px";
-				element.style.position = "absolute";
-				element.style.width = item.clientWidth + 'px';
-			}
-		};
-
-		const generateDragImage = () => {
-			const item = document.querySelectorAll(`.${css.draggableItem}`)[0];
-			if (item) {
-				singleItemDragContainer = document.createElement("div");
-				setCommonElementStyles(singleItemDragContainer);
-				singleItemDragContainer.style.bottom = "0px";
-				singleItemDragContainer.style.left = "0px";
-				singleItemDragContainer.style.zIndex = '-100';
-				document.body.appendChild(singleItemDragContainer);
-
-				multipleItemDragContainer = document.createElement("div");
-				setCommonElementStyles(multipleItemDragContainer);
-				multipleItemDragContainer.style.height = 1.6 * item.clientHeight + 'px';
-				multipleItemDragContainer.style.top = "0px";
-				multipleItemDragContainer.style.zIndex = '-110';
-				document.body.appendChild(multipleItemDragContainer);
-
-				let div2 = document.createElement("div");
-				setCommonElementStyles(div2);
-				div2.style.top = "0px";
-				div2.style.zIndex = '-100';
-
-				let div3 = document.createElement("div");
-				setCommonElementStyles(div3);
-				div3.style.top = 0.3 * item.clientHeight + 'px';
-				div3.style.zIndex = '-90';
-
-				let div4 = document.createElement("div");
-				setCommonElementStyles(div4);
-				div4.style.top = 0.6 * item.clientHeight + 'px';
-				div4.style.zIndex = '-80';
-
-				multipleItemDragContainer.appendChild(div2);
-				multipleItemDragContainer.appendChild(div3);
-				multipleItemDragContainer.appendChild(div4);
-			}
-		};
-
-		useEffect(() => {
-			generateDragImage();
-		});
 
 		const dropEventListenerFunction = useCallback((ev) => {
 			let element;
@@ -501,9 +449,9 @@ const TransferListBase = kind({
 			ev.dataTransfer.effectAllowed = 'move';
 
 			if (selectedItems.length > 1) {
-				ev.dataTransfer.setDragImage(multipleItemDragContainer, 0, 0);
+				ev.dataTransfer.setDragImage(customImage?.multipleItemDragContainer, 0, 0);
 			} else {
-				ev.dataTransfer.setDragImage(singleItemDragContainer, 0, 0);
+				ev.dataTransfer.setDragImage(customImage?.singleItemDragContainer, 0, 0);
 			}
 		}, [selectedItems.length]);
